requirements.txt
python-telegram-bot==20.0a0
requests==2.31.0
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
apscheduler==3.10.4
